Feat: [BE] 영양제 상세 정보 API 리뷰 데이터 반환 추가
영양제 상세 정보 API 리뷰 데이터 반환 추가

+++ backend/src/main/java/planeat/database/repository/NutrientReviewRepository.java
@@ -12,5 +12,5 @@ public interface NutrientReviewRepository extends JpaRepository<NutrientReview, 
     
     @EntityGraph(attributePaths = {"nutrient"})
     @Query("select r from NutrientReview r where r.nutrient.id = :nutrientId")
-    List<NutrientReview> findAllByNutrientId (@Param("nutrientId") Integer nutrientId);
+    List<NutrientReview> findAllByNutrientId (@Param("nutrientId") Long nutrientId);
 }
